pyproject.toml
[tool.poetry]
name = "django-repeat"
version = "0.1.0"
description = ""
authors = ["enterspacebs <anotherhell48@gmail.com>"]
readme = "README.md"
packages = [{include = "django_repeat"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
